+++ NEWS.md
@@ -8,6 +8,8 @@ editor_options:
 
 -   `PAR()` was previously defined as a wavelength range, which is correct only for computations of photon irradiance. For computation of energy irradiance the correct approach is to use a weighting function based on the energy per photon. This now
 fixed. The previous behaviour of `PAR()` can now be obtained with the new function `PhR()`or by passing an argument to parameter `std`. This change affects only computations of energy irradiance, and although not code breaking, returned values will vary slightly.
+-   `PAR()` gains definitions of historical interest as well as the recently
+proposed extended photosynthetically active radiation, ePAR.
 -   `Plant_bands()` now accepts `"Sellaro"` and `"sensoryBroad"` as argument for `std`.
 -   `Blue()` and `Green` now accept `"broad"` as argument for `std`, corresponding to very broad ranges of wavelengths as frequently used in plant sciences.
 -    Update `UVI()` so that it optionally returns integer values as recommended by WWO for communication of UVI to the geenral public. Default behaviour remains unchanged.